refactor: Remove ddu property from UiActionArguments
