+++ denops/ddu/base/ui.ts
@@ -3,14 +3,11 @@ import type {
   Context,
   DduItem,
   DduOptions,
-  PreviewContext,
-  Previewer,
   SourceInfo,
   UiActionCallback,
   UiOptions,
 } from "../types.ts";
 import type { Denops } from "jsr:@denops/std@~7.0.3";
-import type { Ddu } from "../ddu.ts";
 
 export type UiActions<Params extends BaseParams> = Record<
   string,
@@ -113,22 +110,6 @@ export type UpdateCursorArguments<Params extends BaseParams> = {
   options: DduOptions;
   uiOptions: UiOptions;
   uiParams: Params;
-};
-
-export type UiActionArguments<Params extends BaseParams> = {
-  denops: Denops;
-  ddu: Ddu;
-  context: Context;
-  options: DduOptions;
-  uiOptions: UiOptions;
-  uiParams: Params;
-  actionParams: unknown;
-  getPreviewer?: (
-    denops: Denops,
-    item: DduItem,
-    actionParams: BaseParams,
-    previewContext: PreviewContext,
-  ) => Promise<Previewer | undefined>;
 };
 
 export abstract class BaseUi<Params extends BaseParams> {
