+++ denops/ddu/ddu.ts
@@ -931,7 +931,6 @@ export class Ddu {
     } else {
       ret = await action({
         denops,
-        ddu: this,
         context: this.#context,
         options: this.#options,
         uiOptions,
